[fix] kafka 역직렬화 수정

See merge request s11-bigdata-recom-sub1/S11P21E206!97

+++ backend/logger/src/main/java/com/flicker/logger/dto/MovieAverageRating.java
@@ -7,12 +7,12 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 public class MovieAverageRating {
-    private Long movieSeq;
-    private Long reviewSeq;
+    private Integer movieSeq;
+    private Integer reviewSeq;
     private Integer movieCount;
     private Double movieTotalRating;
 
-    public MovieAverageRating(Long movieSeq, Integer movieCount, Double movieTotalRating) {
+    public MovieAverageRating(Integer movieSeq, Integer movieCount, Double movieTotalRating) {
         this.movieSeq = movieSeq;
         this.movieCount = movieCount;
         this.movieTotalRating = movieTotalRating;
